requirements.txt
numpy
PyYAML
requests
torch
tqdm
